improve command and disboard code flow
