+++ src/commands/Bump reminder/setbumpmessage.js
@@ -3,11 +3,10 @@ module.exports = {
         if (message.content) return message.channel.createMessage( {"embed": {"title": `Error`,"description":"Improper syntax.", "color": 5747894}}).catch(err => console.error("Cannot send messages to this channel", err));
         if (!message.member.permission.has("administrator")) return message.channel.createMessage( {"embed": {"title": `Error`,"description":"You must be the owner of this server or an administrator.", "color": 5747894}}).catch(err => console.error("Cannot send messages to this channel", err));
 
-        const guild = await sharder.query({text: 'SELECT * FROM guilds WHERE serverid = $1', values: [message.channel.guild.id]});
-        
+        const guild = await sharder.ipc.command("db", {text: 'SELECT * FROM guilds WHERE serverid = $1', values: [message.channel.guild.id]}, true);
         if (guild.length === 0) return message.channel.createMessage( {"embed": {"title": `Error`,"description":"Make a valid bump first", "color": 5747894}}).catch(err => console.error("Cannot send messages to this channel", err));
 
-        sharder.query({text: 'UPDATE guilds SET bumpmessage = $1  WHERE serverid = $2', values: [message.content, message.channel.guild.id]});
+        sharder.ipc.command("db", {text: 'UPDATE guilds SET bumpmessage = $1  WHERE serverid = $2', values: [message.content, message.channel.guild.id]}, true);
         message.channel.createMessage( {"embed": {"title": `Done`,"description":"The new reminder message for this server is: `"+message.content+"`", "color": 5747894}}).catch(err => console.error("Cannot send messages to this channel", err));
     },
     "help":[
